webapp/requirements.txt
Flask==3.0.2
flask_cors==4.0.0
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-storage-blob==12.19.1
opencensus-ext-azure==1.1.13
azure-search-documents==11.4.0